refactor runner print methods, relocate runner/status types
