+++ runner/status.go
@@ -1,5 +1,98 @@
 package runner
 
+import (
+	"bytes"
+	"fmt"
+)
+
+type RunId string
+type ProcessState int
+
+const (
+	// An unambiguous 0-value.
+	UNKNOWN ProcessState = iota
+	// Waiting to run.
+	PENDING
+	// Preparing to run (e.g., checking out the Snapshot)
+	PREPARING
+	// Running
+	RUNNING
+
+	// States below are end states
+	// a Process in an end state will not change its state
+
+	// Succeeded or failed yielding an exit code. Only state with an exit code.
+	COMPLETE
+	// Run mechanism failed and run is no longer active. Retry may or may not work.
+	FAILED
+	// User requested that the run be killed.
+	ABORTED
+	// Operation timed out and was killed.
+	TIMEDOUT
+	// Invalid or error'd request. Original runner state not affected. Retry may work after mutation.
+	BADREQUEST
+)
+
+func (p ProcessState) IsDone() bool {
+	return p == COMPLETE || p == FAILED || p == ABORTED || p == TIMEDOUT
+}
+
+func (p ProcessState) String() string {
+	switch p {
+	case UNKNOWN:
+		return "UNKNOWN"
+	case PENDING:
+		return "PENDING"
+	case PREPARING:
+		return "PREPARING"
+	case RUNNING:
+		return "RUNNING"
+	case COMPLETE:
+		return "COMPLETE"
+	case FAILED:
+		return "FAILED"
+	case ABORTED:
+		return "ABORTED"
+	case TIMEDOUT:
+		return "TIMEDOUT"
+	case BADREQUEST:
+		return "BADREQUEST"
+	default:
+		panic(fmt.Sprintf("Unexpected ProcessState %v", int(p)))
+	}
+}
+
+// Returned by the coordinator when a run request is made.
+type ProcessStatus struct {
+	RunId RunId
+	State ProcessState
+	// References to stdout and stderr, not their text
+	StdoutRef string
+	StderrRef string
+
+	// Only valid if State == COMPLETE
+	ExitCode int
+
+	// Only valid if State == (FAILED || BADREQUEST)
+	Error string
+}
+
+func (p ProcessStatus) String() string {
+	var b bytes.Buffer
+	b.WriteString(fmt.Sprintf("--- Process Status ---\n\tID:\t\t%s\n\tState:\t\t%s\n", p.RunId, p.State))
+
+	if p.State == COMPLETE {
+		b.WriteString(fmt.Sprintf("\tExitCode:\t%d\n", p.ExitCode))
+	}
+	if p.State == FAILED || p.State == BADREQUEST {
+		b.WriteString(fmt.Sprintf("\tError:\t\t%s\n", p.Error))
+	}
+
+	b.WriteString(fmt.Sprintf("\tStdout:\t\t%s\n\tStderr:\t\t%s\n", p.StdoutRef, p.StderrRef))
+
+	return b.String()
+}
+
 // Helper functions to create ProcessStatus
 
 func AbortStatus(runId RunId) (r ProcessStatus) {
